requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
requests==2.25.1
flasK 
flask-cors 
google-generativeai
Werkzeug==2.0.1
gunicorn==20.1.0